// src/lib/bgm/schema.d.ts
export interface components {
    schemas: {
// ...
        Episode: {
            /** ID */
            id: number;
            /**
             * Type
             * @description `0` 本篇，`1` SP，`2` OP，`3` ED
             */
            type: number;
            /** Name */
            name: string;
            /** Name Cn */
            name_cn: string;
            /**
             * Sort
             * @description 同类条目的排序和集数
             */
            sort: number;
            /**
             * Ep
             * @description 条目内的集数, 从`1`开始。非本篇剧集的此字段无意义
             */
            ep?: number;
            /** Airdate */
            airdate: string;
            /** Comment */
            comment: number;
            /**
             * Duration
             * @description 维基人填写的原始时长
             */
            duration: string;
            /**
             * Desc
             * @description 简介
             */
            desc: string;
            /**
             * Disc
             * @description 音乐曲目的碟片数
             */
            disc: number;
            /** @description 服务器解析的时长，无法解析时为 `0` */
            duration_seconds?: number;
        };
// ...
        UserSubjectCollection: {
            /**
             * Subject ID
             * @example 8
             */
            subject_id: number;
            subject_type: components["schemas"]["SubjectType"];
            /**
             * Rate
             * @example 4
             */
            rate: number;
            type: components["schemas"]["SubjectCollectionType"];
            /**
             * Comment
             * @example 看看
             */
            comment?: string;
            /**
             * Tags
             * @example [
             *       "柯南",
             *       "万年小学生",
             *       "推理",
             *       "青山刚昌",
             *       "TV"
             *     ]
             */
            tags: string[];
            /**
             * Ep Status
             * @example 5
             */
            ep_status: number;
            /**
             * Vol Status
             * @example 0
             */
            vol_status: number;
            /**
             * Updated At
             * Format: date-time
             * @description 本时间并不代表条目的收藏时间。修改评分，评价，章节观看状态等收藏信息时未更新此时间是一个 bug。请不要依赖此特性
             * @example 2022-06-19T18:44:13.6140127+08:00
             */
            updated_at: string;
            /** Private */
            private: boolean;
            subject?: components["schemas"]["SlimSubject"];
        };
// ...
    };
    responses: {
        /** @description Bad Request */
        400: {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["ErrorDetail"];
            };
        };
        /** @description Unauthorized */
        401: {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["ErrorDetail"];
            };
        };
        /** @description Not Found */
        404: {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["ErrorDetail"];
            };
        };
        /** @description Internal Server Error */
        500: {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["ErrorDetail"];
            };
        };
        /** @description Successful Response */
        "200-no-content": {
            headers: {
                [name: string]: unknown;
            };
            content?: never;
        };
    };
// ...
}
export type $defs = Record<string, never>;
export interface operations {
    searchSubjects: {
        parameters: {
            query?: {
                /** @description 分页参数 */
                limit?: number;
                /** @description 分页参数 */
                offset?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: {
            content: {
                "application/json": {
                    keyword: string;
                    /**
                     * @description 排序规则
                     *
                     *     - `match` meilisearch 的默认排序，按照匹配程度
                     *     - `heat` 收藏人数
                     *     - `rank` 排名由高到低
                     *     - `score` 评分
                     * @default match
                     * @example rank
                     * @enum {string}
                     */
                    sort?: "match" | "heat" | "rank" | "score";
                    /** @description 不同条件之间是 `且` 的关系 */
                    filter?: {
                        /** @description 条目类型，参照 `SubjectType` enum，多值之间为 `或` 的关系。 */
                        type?: components["schemas"]["SubjectType"][];
                        /**
                         * @description 公共标签。多个值之间为 `且` 关系。可以用 `-` 排除标签。比如 `-科幻` 可以排除科幻标签。
                         * @example [
                         *       "童年",
                         *       "原创"
                         *     ]
                         */
                        meta_tags?: string[];
                        /**
                         * @description 标签，可以多次出现。多值之间为 `且` 关系。
                         * @example [
                         *       "童年",
                         *       "原创"
                         *     ]
                         */
                        tag?: string[];
                        /**
                         * @description 播出日期/发售日期，日期必需为 `YYYY-MM-DD` 格式。多值之间为 `且` 关系。
                         * @example [
                         *       ">=2020-07-01",
                         *       "<2020-10-01"
                         *     ]
                         */
                        air_date?: string[];
                        /**
                         * @description 用于搜索指定评分的条目，多值之间为 `且` 关系。
                         * @example [
                         *       ">=6",
                         *       "<8"
                         *     ]
                         */
                        rating?: string[];
                        /**
                         * @description 用于按照评分人数筛选条目，多值之间为 `且` 关系，格式与 `rating` 相同。
                         * @example [
                         *       ">=200",
                         *       "<5000"
                         *     ]
                         */
                        rating_count?: string[];
                        /**
                         * @description 用于搜索指定排名的条目，多值之间为 `且` 关系。
                         * @example [
                         *       ">10",
                         *       "<=18"
                         *     ]
                         */
                        rank?: string[];
                        /**
                         * @description 无权限的用户会直接忽略此字段，不会返回R18条目。
                         *
                         *     默认或者 `null` 会返回包含 R18 的所有搜索结果。
                         *
                         *     `true` 只会返回 R18 条目。
                         *
                         *     `false` 只会返回非 R18 条目。
                         */
                        nsfw?: boolean;
                    };
                };
            };
        };
        responses: {
            /** @description 返回搜索结果 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["Paged_Subject"];
                };
            };
        };
    };
// ...
    getSubjectById: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                /** @description 条目 ID */
                subject_id: components["parameters"]["path_subject_id"];
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Successful Response */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["Subject"];
                };
            };
            /** @description Validation Error */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorDetail"];
                };
            };
            /** @description Not Found */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorDetail"];
                };
            };
        };
    };
// ...
    getEpisodes: {
        parameters: {
            query: {
                /** @description 条目 ID */
                subject_id: components["parameters"]["query_subject_id"];
                /** @description 参照章节的`type` */
                type?: components["schemas"]["EpType"];
                /** @description 分页参数 */
                limit?: number;
                /** @description 分页参数 */
                offset?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Successful Response */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["Paged_Episode"];
                };
            };
            /** @description Validation Error */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorDetail"];
                };
            };
            /** @description Not Found */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorDetail"];
                };
            };
        };
    };
    getEpisodeById: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                /** @description 章节 ID */
                episode_id: components["parameters"]["path_episode_id"];
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Successful Response */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["EpisodeDetail"];
                };
            };
            /** @description Validation Error */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorDetail"];
                };
            };
            /** @description Not Found */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorDetail"];
                };
            };
        };
    };
// ...
}